cppTransclust v1.0, working fpt,force. TODO minor optimizations in FPT (ll BFS, 1D costmatrix)

+++ transclust/include/ClusteringResult.hpp
@@ -4,7 +4,7 @@
 
 // Struct holding a partial clustering
 struct ClusteringResult {
-	float cost;
+	double cost;
 	std::vector<unsigned> membership; 
 };
 #endif
