+++ transclust/include/Result.hpp
@@ -12,8 +12,8 @@ class Result{
 		void dump();
 	private:
 		std::vector<std::string> id2object;
-		std::map<float,float> cost;
-		std::map<float,std::vector<std::vector<unsigned>>> clusters;
+		std::map<double,double> cost;
+		std::map<double,std::vector<std::vector<unsigned>>> clusters;
 };
